requirements.txt
pandas
scikit-learn
nltk
streamlit
sqlite3
pickle
Pillow